Change the way properties are asked for, in preparation of starting to return the default values...

+++ src/model/CokeMentosBottle.h
@@ -179,7 +179,7 @@ private:
 	float thePreviousAngVelocity;
 
 	/// constant that describes how much the coke splatter acts on the bottle
-	qreal theThrust;
+	float theThrust;
 
 
 };
